widget: Support using fzy as file picker
This adds a new possible setting for the corresponding option, which
will use fzy [1] as the file picker:

  zstyle :notes:widget picker fzy

Unfortunately, certain features cannot be supported due to limitations
in fzy, the most noticeable being the lack of preview window, and that
Ctrl-N does not work, as fzy cannot bind keys to custom actions. It is
minimalistic, though!

[1] https://github.com/jhawthorn/fzy